fix: remove tab, panel component and separate the eventfilters
