+++ web/src/App.js
@@ -1,5 +1,6 @@
+
 import React from 'react';
-import Eventslist from './pages/eventslist';
+import Landing from './pages/landing';
 import ApolloClient from "apollo-boost";
 import { ApolloProvider } from "react-apollo";
 import {
@@ -29,7 +30,7 @@ function App() {
                       <Signin/>
                   </Route>
                   <Route path="/">
-                      Main
+                    <Eventlist/>
                   </Route>
               </Switch>
           </Router>
